Defaults the api test helpers for terser usage
I'm still on the fence if I would want to use these beyond the API
layer. On the one hand it is nice to have all tests call all the way
through to the api. On the other, it seems like mocking the API file
itself removes implementation knowledge from downstream tests. For
example, do tests for a component need to be concerned with urls
and status codes?