Switch to using vitest instead of jest
Aside from the docs themselves this was helpful:
https://github.com/vitest-dev/vitest/tree/main/examples/react-testing-lib-msw

The test base is so small in this project that I am sure there are
still things that are not wired up, but this seems like a good start.
MSW is one of the bigger pieces and that seems to be working.