Adds basic loader and error message
This is NOT fully developed at this point, but a better starter than
nothing. This at least gives some ideas as to how you would wire things
up and properties you would want to include. The end goal isn't a "drop
in" set of components but more some guidelines to get you moving in the
right direction.